refactor: Move "domain" into config.ts
This avoids circular dependencies
