+++ src/mail-service/templates/components/user-link.tsx
@@ -1,4 +1,4 @@
-import { domain } from '..'
+import { domain } from '../../../config'
 
 export function UserLink({
   user,
