+++ src/mail-service/templates/index.tsx
@@ -1,3 +1,1 @@
 export * from './components/notification-mail'
-
-export const domain = 'https://serlo.org'
